+++ app/src/main/java/app/dreamjournal/data/dream/DreamTagCrossRef.kt
@@ -4,6 +4,6 @@ import androidx.room.Entity
 
 @Entity(primaryKeys = ["dreamId", "tagId"])
 data class DreamTagCrossRef(
-    val dreamId: Int,
-    val tagId: Int,
+    val dreamId: Long,
+    val tagId: Long,
 )
